app/requirements.txt
asgiref==3.5.2
async-generator==1.10
attrs==22.1.0
bezier==2021.2.12
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
Django==4.1.2
exceptiongroup==1.0.0
fake-useragent==0.1.11
Flask==2.2.2
free-proxy==1.0.6
h11==0.14.0
h2==2.6.2
hpack==3.0.0
hyper==0.7.0
hyperframe==3.2.0
idna==3.4
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
Js2Py==0.71
lxml==4.9.1
MarkupSafe==2.1.1
MouseInfo==0.1.3
numpy==1.23.4
outcome==1.2.0
pyaes==1.6.1
pyasn1==0.4.8
PyAutoGUI==0.9.53
pycparser==2.21
pydantic==1.10.2
PyGetWindow==0.0.9
pyjsparser==2.7.1
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
PySocks==1.7.1
pytweening==1.0.4
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.28.1
rsa==4.9
selenium==4.5.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.3
Telethon==1.25.4
trio==0.22.0
trio-websocket==0.9.2
tzdata==2022.5
tzlocal==4.2
urllib3==1.26.12
useragent==0.1.1
Werkzeug==2.2.2
wsproto==1.2.0
zipp==3.10.0
